Warn about Sqlite3 version only when used
